+++ src/app/features/entities/entity-view/entity-view.component.ts
@@ -3,6 +3,8 @@ import { ActivatedRoute, Router } from '@angular/router';
 
 import { Observable, Subscription } from 'rxjs';
 import { tap } from 'rxjs/operators';
+import { CalendarEvent } from 'angular-calendar';
+
 import Swal from 'sweetalert2';
 
 import { BaseType } from '@models/base';
@@ -12,7 +14,9 @@ import { AuthService } from '@auth/auth.service';
 import { SeoService } from '@services/seo.service';
 import { ItemSocialService } from '@services/items-social.service';
 import { EntityService } from '@services/entities.service';
+import { EventService } from '@services/events.service';
 import { UserService } from '@services/users.service';
+import { IEvent, Event } from '@models/event';
 
 @Component({
   selector: 'app-entity-view',
@@ -22,12 +26,23 @@ import { UserService } from '@services/users.service';
 export class EntityViewComponent implements OnInit, OnDestroy {
 
   public entity$: Observable<IEntity | undefined> | null = null;
+  events$: Observable<IEvent[]>;
   public idEntity: string;
   public entityName: string;
   public userLogged: IUser;
 
   public BTN_IMG_FAVORITE_ON = 'assets/svg/favorite-on.svg';
   public BTN_IMG_FAVORITE_OFF = 'assets/svg/favorite-off.svg';
+
+  readonly today = new Date();
+  readonly DATE_MIN = new Date(
+      this.today.getFullYear()-1,
+      this.today.getMonth(),
+      this.today.getDay()).toISOString().substr(0, 10);
+  readonly DATE_MAX = new Date(
+      this.today.getFullYear()+1,
+      this.today.getMonth(),
+      this.today.getDay()).toISOString().substr(0, 10);
 
   private listOfObservers: Array<Subscription> = [];
 
@@ -36,6 +51,7 @@ export class EntityViewComponent implements OnInit, OnDestroy {
     private router: Router,
     private seo: SeoService,
     private entitiesSrv: EntityService,
+    private eventsSrv: EventService,
     private itemSocialSrv: ItemSocialService,
     public authSvc: AuthService,
     private userSrv: UserService,
@@ -55,6 +71,7 @@ export class EntityViewComponent implements OnInit, OnDestroy {
     this.idEntity = this.route.snapshot.paramMap.get('id');
     if ( this.idEntity ) {
       this.getDetails(this.idEntity);
+      this.fetchEvents();
     }
 
     this.listOfObservers.push( subs1$ );
@@ -72,6 +89,13 @@ export class EntityViewComponent implements OnInit, OnDestroy {
             });
           })
       );
+  }
+
+  fetchEvents(): void {
+    this.events$ = this.eventsSrv.getEventsByEntityAndRange(
+        this.DATE_MIN, this.DATE_MAX,
+        this.idEntity
+    );
   }
 
   public setEntityFav(isFav: boolean): void {
@@ -106,6 +130,10 @@ export class EntityViewComponent implements OnInit, OnDestroy {
 
   public editItem(): void {
     this.router.navigate([`/${Entity.PATH_URL}/${this.idEntity}/editar`]);
+  }
+
+  public clickItem(event: IEvent): void {
+      this.router.navigate([`/${Event.PATH_URL}/${event.id}`]);
   }
 
   ngOnDestroy(): void {