+++ src/environments/environment.prod.ts
@@ -1,7 +1,7 @@
 export const environment = {
   production: true,
-  appVersion: '2.0.2.050',
-  releaseDate: '2022-06-23 00:00',
+  appVersion: '2.0.2.051',
+  releaseDate: '2022-06-27 00:00',
   baseUrl: 'https://agendarinconera.es',
   setAudit: true,
   storiesNDaysAhead: 7,