+++ changelog.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `feature` Add social event audit.
 - `feature` Enable entity-favorite from event-views.
 - `feature` Enable entity-favorite from entity-view.
+- `feature` List events in entity-view.
 
 ### Fix